small project refactor
