+++ .gitignore
@@ -21,3 +21,4 @@
 node_modules
 bower_components
 npm-debug.log
+/jobs/Backend/Task/.vs
