+++ jobs/Backend/Task/ExchangeRateUpdater.Core/IoC/IoCExtensions.cs
@@ -1,0 +1,10 @@
+﻿using ExchangeRateUpdater.Core.Providers;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace ExchangeRateUpdater.Core.IoC;
+
+public static class IoCExtensions
+{
+	public static IServiceCollection RegisterCore(this IServiceCollection services) =>
+		services.AddTransient<IExchangeRateProvider, ExchangeRateProvider>();
+}
