+++ jobs/Backend/Task/ExchangeRateUpdater.Core/Models/Currency.cs
@@ -1,0 +1,20 @@
+﻿namespace ExchangeRateUpdater.Core.Models
+{
+	public class Currency
+	{
+		public Currency(string code)
+		{
+			Code = code;
+		}
+
+		/// <summary>
+		/// Three-letter ISO 4217 code of the currency.
+		/// </summary>
+		public string Code { get; }
+
+		public override string ToString()
+		{
+			return Code;
+		}
+	}
+}
