+++ jobs/Backend/Task/ExchangeRateUpdater.Core/Providers/ExchangeRateProvider.cs
@@ -1,0 +1,11 @@
+﻿using ExchangeRateUpdater.Core.Models;
+
+namespace ExchangeRateUpdater.Core.Providers;
+
+internal class ExchangeRateProvider : IExchangeRateProvider
+{
+	public IEnumerable<ExchangeRate> GetExchangeRates(IEnumerable<Currency> currencies)
+	{
+		return Enumerable.Empty<ExchangeRate>();
+	}
+}
