+++ jobs/Backend/Task/ExchangeRateUpdater.sln
@@ -1,9 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34018.315
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater", "ExchangeRateUpdater.csproj", "{7B2695D6-D24C-4460-A58E-A10F08550CE0}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1.Presentation", "1.Presentation", "{4AA0475C-D8D6-4323-B640-A3CF68BF5CA3}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "2.Core", "2.Core", "{2CDCDD1F-68F0-41F1-8CDF-4256EC6DADC9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater.Console", "ExchangeRateUpdater.Console\ExchangeRateUpdater.Console.csproj", "{42B3A6EC-12C9-4E99-8588-FC57E4E94D8B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater.Core", "ExchangeRateUpdater.Core\ExchangeRateUpdater.Core.csproj", "{A1BB4374-C2B3-4CF9-96F4-0F6A1CB3020E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,12 +17,23 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42B3A6EC-12C9-4E99-8588-FC57E4E94D8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42B3A6EC-12C9-4E99-8588-FC57E4E94D8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42B3A6EC-12C9-4E99-8588-FC57E4E94D8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42B3A6EC-12C9-4E99-8588-FC57E4E94D8B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1BB4374-C2B3-4CF9-96F4-0F6A1CB3020E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1BB4374-C2B3-4CF9-96F4-0F6A1CB3020E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1BB4374-C2B3-4CF9-96F4-0F6A1CB3020E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1BB4374-C2B3-4CF9-96F4-0F6A1CB3020E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{42B3A6EC-12C9-4E99-8588-FC57E4E94D8B} = {4AA0475C-D8D6-4323-B640-A3CF68BF5CA3}
+		{A1BB4374-C2B3-4CF9-96F4-0F6A1CB3020E} = {2CDCDD1F-68F0-41F1-8CDF-4256EC6DADC9}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1CF73546-BF48-45D8-BBF0-782C0A2D78BD}
 	EndGlobalSection
 EndGlobal
